+++ include/Node/NodeUtil.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <memory>
 #include "Node/NodeTypes.hpp"
 
@@ -6,7 +7,7 @@ namespace yoctocc {
 
 struct Token;
 
-std::unique_ptr<Node> createNumberNode(const Token* token, int value);
+std::unique_ptr<Node> createNumberNode(const Token* token, int64_t value);
 std::unique_ptr<Node> createUnaryNode(NodeType type, const Token* token, std::unique_ptr<Node> operand);
 std::unique_ptr<Node> createBinaryNode(NodeType type, const Token* token, std::unique_ptr<Node> left, std::unique_ptr<Node> right);
 std::unique_ptr<Node> createVariableNode(const Token* token, Object* variable);
